@@ -1,8 +1,9 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.3.61'
+    ext.kotlin_version = '1.5.10'
     repositories {
+        mavenCentral()
         google()
         jcenter()
         maven {
@@ -12,13 +13,13 @@ buildscript {
     }
     dependencies {
 //        classpath 'com.android.tools.build:gradle:3.2.0'
-        classpath 'com.android.tools.build:gradle:3.5.3'
+        classpath 'com.android.tools.build:gradle:4.2.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-
+//        classpath 'com.hujiang.aspectjx:gradle-android-plugin-aspectjx:2.0.8'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
-        classpath 'com.novoda:bintray-release:0.9'
-        classpath 'com.hujiang.aspectjx:gradle-android-plugin-aspectjx:2.0.4'
+//        classpath 'com.novoda:bintray-release:0.9'
+
     }
 }
 
@@ -31,20 +32,14 @@ allprojects {
 
 allprojects {
     repositories {
+        mavenCentral()
         google()
         jcenter()
         maven{ url 'http://maven.aliyun.com/nexus/content/groups/public/'}
         maven { url "https://raw.githubusercontent.com/Pgyer/mvn_repo_pgyer/master" }
         maven { url "https://jitpack.io" }
-    }
-    tasks.withType(Javadoc) {
-        options{ encoding "UTF-8"
-            charSet 'UTF-8'
-            links "http://docs.oracle.com/javase/7/docs/api"
-        }
-    }
 
-    tasks.getByPath(":core:javadocRelease").enabled = false
+    }
 }
 
 task clean(type: Delete) {
