@@ -5,8 +5,8 @@ package cn.com.heaton.blelibrary.ble.model;
  */
 import android.os.Build;
 import android.os.ParcelUuid;
-import android.support.annotation.Nullable;
-import android.support.annotation.RequiresApi;
+import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 import android.util.ArrayMap;
 import android.util.SparseArray;
 
