@@ -1,6 +1,6 @@
 package com.example.admin.mybledemo;
 
-import android.support.v4.content.FileProvider;
+import androidx.core.content.FileProvider;
 
 public class AppFileProvider extends FileProvider {
 }
