@@ -4,25 +4,26 @@ import android.bluetooth.BluetoothGatt;
 import android.bluetooth.BluetoothGattCharacteristic;
 import android.bluetooth.BluetoothGattService;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.os.HandlerCompat;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.DividerItemDecoration;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.appcompat.app.ActionBar;
 import android.util.Log;
 import android.view.MenuItem;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.DividerItemDecoration;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.admin.mybledemo.R;
 import com.example.admin.mybledemo.Utils;
 import com.example.admin.mybledemo.adapter.DeviceInfoAdapter;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
 import cn.com.heaton.blelibrary.ble.Ble;
-import cn.com.heaton.blelibrary.ble.BleHandler;
 import cn.com.heaton.blelibrary.ble.BleLog;
 import cn.com.heaton.blelibrary.ble.callback.BleConnectCallback;
 import cn.com.heaton.blelibrary.ble.callback.BleNotifyCallback;
@@ -31,55 +32,23 @@ import cn.com.heaton.blelibrary.ble.utils.ByteUtils;
 
 public class DeviceInfoActivity extends AppCompatActivity {
 
-    private static final String TAG = "DeviceInfoActivity";
     public static final String EXTRA_TAG = "device";
+    private static final String TAG = "DeviceInfoActivity";
     private BleDevice bleDevice;
     private Ble<BleDevice> ble;
     private ActionBar actionBar;
     private RecyclerView recyclerView;
     private DeviceInfoAdapter adapter;
     private List<BluetoothGattService> gattServices;
-
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_deviceinfo);
-        initView();
-        initData();
-    }
-
-    private void initData() {
-        ble = Ble.getInstance();
-        bleDevice = getIntent().getParcelableExtra(EXTRA_TAG);
-        if (bleDevice == null) return;
-        ble.connect(bleDevice, connectCallback);
-    }
-
-    private void initView() {
-        actionBar = getSupportActionBar();
-        actionBar.setTitle("详情信息");
-        actionBar.setDisplayHomeAsUpEnabled(true);
-
-        recyclerView = findViewById(R.id.recyclerView);
-        gattServices = new ArrayList<>();
-        adapter = new DeviceInfoAdapter(this, gattServices);
-        recyclerView.setLayoutManager(new LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false));
-        recyclerView.addItemDecoration(new DividerItemDecoration(this,DividerItemDecoration.VERTICAL));
-        recyclerView.getItemAnimator().setChangeDuration(300);
-        recyclerView.getItemAnimator().setMoveDuration(300);
-        recyclerView.setAdapter(adapter);
-    }
-
     private BleConnectCallback<BleDevice> connectCallback = new BleConnectCallback<BleDevice>() {
         @Override
         public void onConnectionChanged(BleDevice device) {
-            Log.e(TAG, "onConnectionChanged: " + device.getConnectionState()+Thread.currentThread().getName());
+            Log.e(TAG, "onConnectionChanged: " + device.getConnectionState() + Thread.currentThread().getName());
             if (device.isConnected()) {
                 actionBar.setSubtitle("已连接");
-            }else if (device.isConnecting()){
+            } else if (device.isConnecting()) {
                 actionBar.setSubtitle("连接中...");
-            }
-            else if (device.isDisconnected()){
+            } else if (device.isDisconnected()) {
                 actionBar.setSubtitle("未连接");
             }
         }
@@ -131,11 +100,41 @@ public class DeviceInfoActivity extends AppCompatActivity {
                 @Override
                 public void onNotifySuccess(BleDevice device) {
                     super.onNotifySuccess(device);
-                    BleLog.e(TAG, "onNotifySuccess: "+device.getBleName());
+                    BleLog.e(TAG, "onNotifySuccess: " + device.getBleName());
                 }
             });
         }
     };
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_deviceinfo);
+        initView();
+        initData();
+    }
+
+    private void initData() {
+        ble = Ble.getInstance();
+        bleDevice = getIntent().getParcelableExtra(EXTRA_TAG);
+        if (bleDevice == null) return;
+        ble.connect(bleDevice, connectCallback);
+    }
+
+    private void initView() {
+        actionBar = getSupportActionBar();
+        actionBar.setTitle("详情信息");
+        actionBar.setDisplayHomeAsUpEnabled(true);
+
+        recyclerView = findViewById(R.id.recyclerView);
+        gattServices = new ArrayList<>();
+        adapter = new DeviceInfoAdapter(this, gattServices);
+        recyclerView.setLayoutManager(new LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false));
+        recyclerView.addItemDecoration(new DividerItemDecoration(this, DividerItemDecoration.VERTICAL));
+        recyclerView.getItemAnimator().setChangeDuration(300);
+        recyclerView.getItemAnimator().setMoveDuration(300);
+        recyclerView.setAdapter(adapter);
+    }
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
@@ -151,10 +150,10 @@ public class DeviceInfoActivity extends AppCompatActivity {
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        if (bleDevice != null){
-            if (bleDevice.isConnecting()){
+        if (bleDevice != null) {
+            if (bleDevice.isConnecting()) {
                 ble.cancelConnecting(bleDevice);
-            }else if (bleDevice.isConnected()){
+            } else if (bleDevice.isConnected()) {
                 ble.disconnect(bleDevice);
             }
         }
